% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputdata_data_linkage_finder.R
\name{findLink}
\alias{findLink}
\title{Title}
\usage{
findLink(allObj = NA, allObjMeta = NA, cutoff = 1)
}
\arguments{
\item{allObj}{}

\item{allObjMeta}{}

\item{cutoff}{}
}
\description{
Title
}
