% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_harmonization.R
\name{get_feild_details}
\alias{get_feild_details}
\title{Title}
\usage{
get_feild_details(feild = NULL, feild_type = NULL, datSource = NULL,
  obj = NULL, meta)
}
\arguments{
\item{feild}{}

\item{feild_type}{}

\item{datSource}{}

\item{obj}{}

\item{meta}{}
}
\value{

}
\description{
Title
}
