% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{input_data}
\alias{input_data}
\title{Input Data}
\usage{
input_data(file = NA, dataType = NA, asObj = TRUE, desc = NA, ...)
}
\arguments{
\item{file}{}

\item{dataType}{}

\item{asObj}{}

\item{desc}{}

\item{...}{}
}
\description{
Generic input function for inputing data, helper functions are in the back
}
