% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{input_phyloTree}
\alias{input_phyloTree}
\title{A helper function that reads in tree file data. Expecting either Newick or Nexus format}
\usage{
input_phyloTree(file = NA, desc = NA, sepLabel = NA,
  metadataFile = NULL, asObj = TRUE, ...)
}
\arguments{
\item{...}{}
}
\value{
a phylo tree object
}
\description{
A helper function that reads in tree file data. Expecting either Newick or Nexus format
}
