% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_shiny_app.r
\name{run_epidrive}
\alias{run_epidrive}
\title{Command to run epivis shiny application}
\usage{
run_epidrive()
}
\description{
Command to run epivis shiny application
}
\details{
run a graphical user interface with epivis
}
